CATO-4448: Make note optional for dormant accounts
To get the required behaviour we need a new box which
indicates that the filing has a revaluation reserve
note. When this is true, AC189 needs to be present and
AC190 needs to equal AC76. If it is not then AC190 is
None and valid and AC189 should be optional.